requirements.txt
# Основні бібліотеки
aniso8601==10.0.0
anyio==4.9.0
argon2-cffi==21.3.0
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
celery==5.5.3
celery-types==0.23.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.5
ddt==1.7.2
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.4
django-countries==7.5.1
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
django-elasticsearch-dsl==7.4.0
django-elasticsearch-dsl-drf>=0.22.4,<0.23
drf-spectacular==0.28.0
drf_yasg==1.21.10
djoser==2.3.3
elasticsearch==7.17.12
factory_boy==3.3.3
fastapi==0.116.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
gTTS==2.5.4
gunicorn>=21.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
moviepy==2.2.1
numpy==2.3.1
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
peewee==3.17.9
Pillow==11.2.1
pluggy==1.6.0
proglog==0.1.12
psycopg2-binary==2.9.10
publicsuffix2==2.20191221
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PySocks==1.7.1
Pygments==2.19.2
python-decouple==3.8
python-dotenv==1.1.1
python-json-logger==3.3.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.31.0
requests-oauthlib==2.0.0
selenium==4.10.0
setuptools>=68.0.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sortedcontainers==2.4.0
sqlparse==0.5.3
starlette==0.47.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions>=4.13.2,<4.14.0
tzdata>=2025.2
uvicorn==0.35.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0

# CI/CD tools
flake8>=6.1

# Тестування
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-django==4.11.1

# Для типів та FastAPI
annotated-types==0.7.0

# OAuth
django-allauth==65.10.0


